+++ modules/QuickEncounter.js
@@ -56,6 +56,7 @@
                 - Asks you when you go to run it (already did this)
                 v0.5.3d: Add option to "freeze" captured tokens so that TokenMold doesn't regenerate HP, name, etc
                 (Default is true so that only newly generated tokens are changed )
+12-Oct-2020     v0.6.0: Allow Quick Encounters to use Compendium links
 */
 
 
@@ -306,12 +307,14 @@ export class QuickEncounter {
     static extractQuickEncounter(journalSheet) {
         const journalEntry = journalSheet.entity;
         const mapNote = journalEntry.sceneNote;
+        //0.6 this now potentially includes Compendium links
         const extractedActors = QuickEncounter.extractActors(journalSheet.element);
         const existingTokens =  journalEntry.getFlag(MODULE_NAME, TOKENS_FLAG_KEY);
 
-        //Minimum Quick Encounter has a Journal Entry, and tokens or actors
+        //Minimum Quick Encounter has a Journal Entry, and tokens or actors (or 0.6 Compendium which turns into Actors)
         //If there isn't a map Note we may need to switch scenes
-        if (journalEntry && ((extractedActors && extractedActors.length) || (existingTokens && existingTokens.length))) {
+        if (journalEntry && ((extractedActors && extractedActors.length) || (existingTokens && existingTokens.length))
+            ) {
             const quickEncounter = {
                 journalEntry : journalEntry,
                 mapNote : mapNote,
@@ -325,6 +328,7 @@ export class QuickEncounter {
     }
 
     static extractActors(html) {
+        const ACTOR = "Actor";
         const entityLinks = html.find(".entity-link");
         if (!entityLinks || !entityLinks.length) {return null;}
 
@@ -332,15 +336,22 @@ export class QuickEncounter {
         const reg = "([0-9]+)[^0-9]*$"; //Matches last "number followed by non-number at the end of a string"
         entityLinks.each((i, el) => {
             const element = $(el);
-            if (element.attr("data-entity") === "Actor") {
-                const dataID = element.attr("data-id");
+            const dataEntity = element.attr("data-entity");
+            const dataID = element.attr("data-id");
+            //0.6 If it's a Compendium we just have a data.pack attribute
+            const dataPackName = element.attr("data-pack"); //Not used if Actor
+            const dataLookup = element.attr("data-lookup");
+            //Get the dataPack entity type (has to be Actor)
+            const dataPack = game.packs.get(dataPackName);
+            if ((dataEntity === ACTOR) || (dataPack && (dataPack.entity === ACTOR))) {
                 const dataName = element.text();
                 const prevSibling = element[0].previousSibling;
                 const possibleInts = prevSibling ? prevSibling.textContent.match(reg) : ["1"];
                 const numActors = parseInt(possibleInts ? possibleInts[0] : "1");
                 extractedActors.push({
                     numActors : numActors ? numActors : 1,
-                    actorID : dataID,
+                    dataPackName : dataPackName,                    //if non-null then this is a Compendium reference
+                    actorID : dataPackName ? dataLookup : dataID, //If Compendium this is a reference into that
                     name : dataName
                 });
             }
@@ -521,8 +532,19 @@ export class QuickEncounter {
         let expandedTokenData = [];
         for (let eActor of extractedActors) {
             let numActors = eActor.numActors;
-            const actor = game.actors.get(eActor.actorID);
-
+            let actor = null;
+            //v0.6 Need to check whether this is a direct Actor reference or from a Compendium
+            if (eActor.dataPackName) {
+                const actorPack = game.packs.get(eActor.dataPackName);
+                if (!actorPack) {continue;}
+                //Import this actor because otherwise you won't be able to see character sheet etc.
+                actor = await game.actors.importFromCollection(eActor.dataPackName, eActor.actorID, {}, {renderSheet: false});
+                //Try This
+                eActor.actorID = actor.id;
+            } else {
+                actor = game.actors.get(eActor.actorID);
+            }
+            if (!actor) {continue;}     //possibly will happen with Compemdium
              //If numActors didn't convert then just create 1 token
              if (!numActors) {numActors = 1;}
 
@@ -536,7 +558,8 @@ export class QuickEncounter {
                  }
                  //Use the prototype token from the Actors
                  tokenData = mergeObject(actor.data.token, tokenData, {inplace: false});
-
+                 //If from a Compendium, we remember that and the original Compendium actorID
+                 if (eActor.dataPackName) {tokenData.compendiumActorID = eActor.actorID;}
                  expandedTokenData.push(tokenData);
              }
         }
@@ -576,13 +599,14 @@ export class QuickEncounter {
             token.control({releaseOthers : false, updateSight : false});
         }
 
-        const tabApp = ui.combat;
-        tabApp.renderPopout(tabApp);
-        //If the tokens are not on the Scene then add them
-
         //Load the recovered tokens into the combat Tracker
         //Only have to toggle one of them to add all the controlled tokens
         await createdTokens[0].toggleCombat();
+
+        //0.6: Moved after toggling combat in case that actually creates the combat entity
+        const tabApp = ui.combat;
+        tabApp.renderPopout(tabApp);
+        //If the tokens are not on the Scene then add them
 
         //Now release control of them as a group, because otherwise the stack is hard to see
         for (const token of createdTokens) {